requirements.txt
blinker==1.9.0
click==8.2.1
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec==2025.7.0
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
python-bidi==0.6.6
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
triton==3.3.1
typing_extensions==4.14.1
Werkzeug==3.1.3
